# MySiteSolution/RestApiTest/my-swagger.json
{
  "x-generator": "NSwag v13.18.0.0 (NJsonSchema v10.8.0.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:44341",
  "schemes": [
    "https"
  ],
  "produces": [
          "application/json"
        ],
  "paths": {
    "/api/v1/categories": {
      "get": {
        "tags": [
          "CategoriesApi"
        ],
        "operationId": "CategoriesApi_GetCategories",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "tags": [
          "CategoriesApi"
        ],
        "operationId": "CategoriesApi_GetCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/api/v1/categories/image/{id}": {
      "get": {
        "tags": [
          "CategoriesApi"
        ],
        "operationId": "CategoriesApi_GetCategoryImage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string",
              "format": "byte"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoriesApi"
        ],
        "operationId": "CategoriesApi_PostCategoryImage",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "image",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string",
              "format": "byte"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "ProductsApi"
        ],
        "operationId": "ProductsApi_GetProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "limit",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductsApi"
        ],
        "operationId": "ProductsApi_PostProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "ProductsApi"
        ],
        "operationId": "ProductsApi_GetProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductsApi"
        ],
        "operationId": "ProductsApi_PutProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProductsApi"
        ],
        "operationId": "ProductsApi_DeleteProduct",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Category": {
      "type": "object",
      "required": [
        "categoryId"
      ],
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int32"
        },
        "categoryName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "productId",
        "categoryId",
        "unitPrice",
        "unitsInStock",
        "unitsOnOrder",
        "reorderLevel",
        "discontinued"
      ],
      "properties": {
        "productId": {
          "type": "integer",
          "format": "int32"
        },
        "productName": {
          "type": "string"
        },
        "categoryId": {
          "type": "integer",
          "format": "int32"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "quantityPerUnit": {
          "type": "string"
        },
        "unitPrice": {
          "type": "number",
          "format": "decimal"
        },
        "unitsInStock": {
          "type": "integer"
        },
        "unitsOnOrder": {
          "type": "integer"
        },
        "reorderLevel": {
          "type": "integer"
        },
        "discontinued": {
          "type": "boolean"
        }
      }
    }
  }
}